+++ .gitignore
@@ -1,3 +1,4 @@
 .idea
+*.iml
 Cargo.lock
 target
