+++ src/systems/physics_stepper.rs
@@ -1,6 +1,8 @@
 use std::marker::PhantomData;
 
-use specs::{world::Index, Entities, Entity, Read, System, SystemData, World, Write, WriteExpect};
+use specs::{
+    world::Index, Entities, Entity, Read, World, System, SystemData, Write, WriteExpect,
+};
 
 use crate::events::{ContactEvent, ContactEvents, ContactType, ProximityEvent, ProximityEvents};
 use nalgebra::RealField;
